ConectaVoz-Web-App/backend/requirements.txt
# Core Django
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0

# Authentication
djangorestframework-simplejwt>=5.2.0
django-allauth>=0.54.0

# API Documentation
drf-spectacular>=0.26.0

# Development Tools
django-extensions>=3.2.0

# Image handling (for user photos)
Pillow>=10.0.0

# Database drivers
# SQLite is included with Python by default
# For production databases, uncomment as needed:
# psycopg2-binary>=2.9.0  # PostgreSQL
# mysqlclient>=2.1.0      # MySQL

# Utility packages
python-dateutil>=2.8.0
pytz>=2023.3

# File handling
whitenoise>=6.5.0  # Static file serving

# Security
django-ratelimit>=4.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0

# Code quality
flake8>=6.0.0
black>=23.0.0